antigravity_v2/requirements.txt
google-genai
ffmpeg-python
moviepy
pexels-api-py
rich
schedule
requests
